+++ changelog.md
@@ -5,6 +5,7 @@
 - New SAO: Mage's Arcane Blast
 - New GAB: Mage's Arcane Missiles, during Arcane Blast
 - New GAB: Mage's Arcane Explosion, during Arcane Blast
+- Updated SAO: Mage's Clearcasting is back to 150% size, to avoid overlap
 
 #### v1.1.0-beta (2023-12-01)